Toml Parsing
Added a function which reads in a toml file, and creates a dictionary
(of dictionaries) for each table in the file. For example, if the
test.toml file contains the following

[config]
path = sources.bib

the resulting dictionary would be

{'config': {'path': 'sources.bib'}}

This will be the way that the user tells the plugin where to find the
.bib file, as well as where they will set any options that I can come up
with.

I also added a keybinding which calls this function (<Leader>r), which
the user can use to reload their configuration file if they change it.