Read Sources File
Added a function which reads the sources file and searches for the
labels that the user has created for their sources. Currently, this
simply echos all the sources, which in itself is a little usefull, since
this is almost what we want, as it allows the user to view a list of all
the defined sources while editing another document.

Finally, this commit also introduces a Run method, which is what will
actually be bound to a command or keystroke. It runs the ParseToml
method as well as the ReadSources method.