Menu Functionality
Added a (global) function to insert a string at the current cursor
position in the current file. Each menu that this plugin creates now
calls this function. The result is that now when the user presses <F4>
(After loading the config file with <Leader>r, this will be changed
eventually), they are presented with a menu of all their sources labels,
and when they select one, it is inserted under the cursor.