Fail Checks
Began making the script a little more robust against errors.  There is
still work to do but I believe that currently any type of file access
error will be handled correctly.  The main things left to do are to
check the syntax of the config and .bib file before adding things to the
config dictionary and menus.