First Config Option
Added support for the offset option (which should be located in the
'config' table), which specifies how many characters to the right or
left of the cursor position the source label should be entered. This is
up to the user, but the two values which make the most sense to me are 0
(the default) which pastes the text directly to the right of the cursor,
and -1, which pastes the test in front of the character under the cursor.