Fixes and Initializing
Fixed the name of the config file, as well as the key bindings which
call the functions of this plugin. The new keybindings are

<Leader>r       To reload the config file, and reread sources file
<Leader>s       To access the menu of sources.

Currently, many things are echoed, and since this script now actually
calls the functions when Vim is started, these echos should be changed
to writing to a log file.