Logging
Added support for logging.  The user may define a log file in the
'config' table using the 'log' key. If the value of this key does not
exist in the config file, the default log filename is
'.sourcesmenu.log'. If there is a problem opening the log file, there is
currently no backup plan in place, it simply fails.