Help File
Added a description of the config file to the help file, as well as a
description of the key mappings that the plugin sets up for the user.
Currently, there is no way to override the key mappings that the script
uses, but this will be changed in a future version of this plugin.