Popup Menu option, and tests
Added an option to use popup menus instead of the wildmenu. This can be
enabled by adding the line 'popup = 1' to the [config] table of the
config file.

In this commit, I also introduce a script to generate some files for
testing purposes, as well as a makefile to clean everything up when
needed.