Checks
Added some checks to verify that things are loaded correctly, as well as
printing some errors to the user when things go wrong, such as invalid
keys, values, or table names in the config file, the log file not being
writable, etc.

This commit also expands the tests that the test script performs, there
is currently a test which contains a working example, as well as one
which loads a bad config file.