Autoload
Separated the ReadSources method from the main file, this is now located
in the autoload directory, meaning that when this plugin is used, this
file will only be parsed when needed. This is a good place for this file
since I plan on extending the filetypes allowed to be read by this
plugin, and only one will be needed at a time.

Currently, there is NOT a config option where the user can specify what
filetype they are using.