Single Place to Read options
Added a method which reads all the options set in the config dictionary.
This makes the code easier to read, since there are fewer (one) places
where the config is checked, and thus every other method which needs
this data simple calls what it needs, since it will be run after the
read options method.