Work from Subdirectories
This allows the config file to be located in a parent directory to the
cwd.  This sometimes happens in projects with many TeX files which may
be located in directories, in this case, the user need only place their
sourcesmenu.toml config file in the top directory, since the plugin now
searches up the directory structure before it moves on to finding global
config files.